refactor: move loading component to root of app
